[Feat.] Add 2.3.3 rule
