+++ public/rulesets/default/2_3_3_Request_body.yaml
@@ -1,42 +1,15 @@
 rules:
   - id: "2.3.3"
     title: "Support JSON/BSON Encapsulation in Request Body"
-    message: "The request body must support both JSON and BSON encapsulation. JSON must conform to key:value format, and BSON-specific types must be correctly declared when 'application/bson' is used."
+    message: "The request body must support both JSON and BSON encapsulation."
     option: Mandatory
     location: paths
-    element:
-      - requestBody
-    call: {
-      function: jsonBsonEncapsulationCheck,
-      functionParams: {
-        "jsonExamples": [
-          {
-            "tenant_id": "t_00001",
-            "disk_format": "none"
-          }
-        ],
-        "acceptedMimeTypes": ["application/json", "application/bson"],
-        "bsonTypes": [
-          {"type": "boolean"},
-          {"type": "integer", "format": "byte"},
-          {"type": "integer", "format": "int32"},
-          {"type": "integer", "format": "int64"},
-          {"type": "integer", "format": "uint64"},
-          {"type": "number", "format": "double"},
-          {"type": "number", "format": "bson-decimal128"},
-          {"type": "string"},
-          {"type": "bson-document"},
-          {"type": "items"},
-          {"type": "string", "format": "bson-binary"},
-          {"type": "string", "format": "objectid"},
-          {"type": "string", "format": "bson-date-time"},
-          {"type": "string", "format": "regex"},
-          {"type": "string", "format": "jsCode"},
-          {"type": "string", "format": "bson-timestamp"},
-          {"type": "string", "format": "minKey"},
-          {"type": "string", "format": "maxKey"}
-        ]
-      }
-    }
+    element: requestBody
+    call:
+      function: requestEncapsulationCheck
+      functionParams:
+        content:
+          - application/json
+          - application/bson
     severity: "high"
-    status: devel
+    status: implemented
