--- public/rulesets/default/2_3_3_Response_body.yaml
@@ -1,17 +1,0 @@
-rules:	
-  - id: "2.3.4"
-    title: "Response Body JSON/BSON Encapsulation"
-    message: "The response body must support JSON or BSON encapsulation. If a body is present, the Content-Type must be application/json;charset=UTF-8 or application/bson."
-    option: Mandatory
-    location: paths
-    element:
-      - responses
-    call: {
-      function: responseBodyEncodingCheck,
-      functionParams: {
-        "headers": ["application/json;charset=UTF-8", "application/bson"],
-        "methods": ["post", "put", "delete"]
-      }
-    }
-    severity: "error"
-    status: devel
