+++ public/rulesets/default/2_3_4_Response_body.yaml
@@ -1,0 +1,22 @@
+rules:	
+  - id: "2.3.4"
+    title: "Response Body JSON/BSON Encapsulation"
+    message: "The response body must support JSON or BSON encapsulation. If a body is present, the Content-Type must be application/json or application/bson."
+    option: Mandatory
+    location: paths
+    element:
+      - responses
+    call:
+      function: responseEncapsulationCheck,
+      functionParams:
+        headers:
+          - Content-Type
+        content:
+          - application/json
+          - application/bson
+        methods:
+          - "post"
+          - "put"
+          - "delete"
+    severity: "error"
+    status: devel
